src/main/java/linh/springframework/recipeapp/converters/CategoryCommandToCategory.java
package linh.springframework.recipeapp.converters;

import linh.springframework.recipeapp.commands.CategoryCommand;
import linh.springframework.recipeapp.domain.Category;
import org.springframework.core.convert.converter.Converter;
import org.springframework.stereotype.Component;

@Component
public class CategoryCommandToCategory implements Converter<CategoryCommand, Category> {
    @Override
    public Category convert(CategoryCommand source) {
        if(source == null){
            return  null;
        }
        final Category category = new Category();
        category.setId(source.getId());
        category.setDescription(source.getDescription());
        return category;
    }

}
